Updated project

+++ backend/database/models.py
@@ -71,7 +71,7 @@ class Alert(Base):
 
 class AlertDeliveryLog(Base):
     __tablename__ = 'alert_delivery_logs'
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     session_id = Column(UUID(as_uuid=True), ForeignKey(EMERGENCY_SESSION_ID_FK))
     alert_id = Column(BigInteger)
     provider_message_id = Column(String(100), nullable=True)
@@ -85,7 +85,7 @@ class AlertDeliveryLog(Base):
 
 class AIEvent(Base):
     __tablename__ = 'ai_events'
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     session_id = Column(UUID(as_uuid=True), ForeignKey(EMERGENCY_SESSION_ID_FK))
     event_type = Column(String(50))
     value = Column(DECIMAL(10, 2))
